netdog: Extract lease parsing functions to module
